refactor(tui): give PendingAction one exhaustive confirm lookup (#417)
The Confirm modal's per-action facts were spread over five matches in three
modules -- key resolution, title/colour, question body, modal body kind, and
background -- each ending in a wildcard. Nothing made a new `PendingAction`
handled anywhere. A forgotten one compiled clean and reached the user as a red
modal titled "Overwrite" asking "Overwrite <download target>? The local file is
replaced", with keys bound to nothing it described.

Five matches are now two, both exhaustive with no `_` arm:

- `build_confirm_vm` carries one row per action: title, border, body and
  background together, with the payload destructured in the arm that uses it.
  `confirm_modal_style` and `view_model::confirm_prompt` are gone, absorbed
  into it; `on_off` moved with them.
- `handle_key_confirm` keeps its per-action arms and drops the wildcard. Its
  `None` arm now says what it means: not on the Confirm screen, which
  `screens::lookup` never routes here.

Shape note: the fn-pointer table used by `screens::lookup` does not fit here.
Its columns would each have to re-destructure the action from `AppState` behind
an unreachable else, four times per row. Returning values from one match keeps
the payload in scope and reads as the same table. The per-keypress clone of the
whole action therefore stays -- buying it back would need a mirror enum, i.e.
one more place a new variant must be handled, which is what this removes.

Tests: the five that called the two deleted functions directly move to the
screen module and are raised to `build_confirm_vm`, which returns title, border
and body together. Added a walk over every action (both create-flow steps)
asserting a non-empty question and at least one resolving key -- the compiler
guarantees a row exists, not that it was filled in. Verified it fails when a
question is blanked.

No user-visible change. ADR-0001 and ADR-0002 untouched: key resolution still
returns `KeyOutcome`.
