+++ docs/agents/architecture.md
@@ -41,6 +41,7 @@ Fields stay `pub`. No `#[serde(default)]` on the struct.
 ## Screen state machine
 
 - **`screens::lookup`** (`@src/tui/screens/mod.rs`, issues #377, #388) is the exhaustive match for the per-screen columns: help topic, wheel step, key guard, VM builder, key handler, navigation, and click selection. It first borrows `self.screen`, then calls an `fn(&mut AppState, …)` pointer: that two-phase pattern is #274's borrow rule. `render_screen_vm` matches `ScreenVm`, not `Screen`. `keymap::for_screen` stays in `keymap.rs` (bindings live there; putting them on the lookup would cycle `screens` ↔ `keymap`). No `ScreenModule` trait: the screen files already are the adapters.
+- **`build_confirm_vm` is the exhaustive lookup for `PendingAction`** (`@src/tui/screens/confirm.rs`, issue #417). One arm per pending action carries its title, border colour, body (`ConfirmModalKind`) and background together; there is no `_` arm, so a new variant that forgets its row fails to compile instead of silently inheriting the overwrite gate's destructive prompt. `confirm_modal_style` and `view_model::confirm_prompt` are gone — they were two of the five wildcard-terminated matches this replaced. `handle_key_confirm` is the second (and last) exhaustive match: it still returns `KeyOutcome` (ADR-0002) and still clones the action, because the arms need `&mut AppState`; its `None` arm means "not on the Confirm screen", which `screens::lookup` never routes here. Presentation is asserted through `build_confirm_vm` only — do not reach past it to a per-fact helper.
 - **`Screen` is `Clone`, not `Copy`** — payload variants own screen-local UI (`@src/tui/mod.rs`, issue #242).
 - **`List` stays a unit tag** — dual-pane selection / filters / sorts are session-global on `AppState` (user story 19).
 - Other variants (`Diff`, `Confirm`, `Preview`, `Help`, `Config`, `Revisions`, `Pins`, `Gists`, `GistDetail`, `Palette`, …) carry payloads (body/scroll/return/origin as needed).
